requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.0
Django==5.0.1
geographiclib==2.0
geopy==2.4.1
idna==3.6
libretranslatepy==2.1.1
lxml==5.1.0
pyTelegramBotAPI==4.15.2
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
translate==3.6.1
urllib3==2.1.0

